requirements.txt
# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# HTTP & Web Scraping
requests==2.31.0
beautifulsoup4==4.12.3
#lxml==5.1.0
tavily-python>=0.5.0

# Task Scheduling
schedule==1.2.1

# Configuration
pyyaml==6.0.1
python-dotenv==1.0.0

# Database
aiosqlite==0.19.0

# Testing
pytest==8.0.0
pytest-asyncio==0.23.3
pytest-cov==4.1.0
httpx==0.26.0  # For testing FastAPI endpoints

# Code Quality
black==24.1.1
flake8==7.0.0
